First (very inclomplete) blush at adding authentise integration
We basically are creating new comm and printer objects to use the
authentise service instead of the base octoprint printer and comm
objects. Eventually this will be integrated in such a way that there
will be an option when connecting to the printer to connect using
authentise instead, which will use these objects. For now I have hard
coded them.