More Authentise integration
This adds the rest of the first "working" version of an Authentise
integration with OctoPrint.  It requires the authentise streaming client
to be running in the background, and currently has a hardcoded
authentise_printer_id.  It also requires a creds.txt file which should
contains an apikey secret for communication with Authentise.

It currently doesn't support printing, only issuing gcode commands.

Command responses are forthcomming.